+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v1.0.14
+ * Add Mega Hack Mobile scroll layer to the known stealer list
+ * Add "alk.better-touch-prio/steals-touch" as a user flag for marking as touch stealer
+
 ## v1.0.13
  * Remove the nested false positives from leak detection